+++ ReleaseNotes.md
@@ -2,11 +2,12 @@ Q. Never mind the fluff -- how do I get started?
 
 A. Start with the README : https://github.com/SteveGilham/altcode.test/blob/master/README.md
 
-# 1.0.xx
+# 2.0.xx
 * [BREAKING] make the `AssertionMatch` type specific to each package to avoid name clashes if multiple test suites are in use
 * [BREAKING] remove support for obsolescing Expecto `Expect.floatEqual`
 * [BREAKING] remove support for obsolescing NUnit `Assert.AreEqual<Nullable<double>>` methods
-* [BREAKING] rename `AltCode.Test.Xunit.AltAssert.DoesNotAssertionMatch` & `AssterionMatches` to `AltCode.Test.Xunit.AltAssert.DoesNotMatch` and `Matches` to line up with the methods being shadowed in `Xunit.Assert`
+* [BREAKING] rename `AltCode.Test.Xunit.AltAssert.DoesNotAssertionMatch` & `AssertionMatches` to `AltCode.Test.Xunit.AltAssert.DoesNotMatch` and `Matches` to line up with the methods being shadowed in `Xunit.Assert`
+* [BREAKING] in NUnit methods with `...,message, args)` correctly declare args as a `params object[]` making the methods take variable numbers of arguments
 * support `netstandard2.0` only (meaning ≥ `net472` compatible)
 * support Expecto ≥ 9.0.4, NUnit ≥ 3.13.3, Xunit ≥ 2.4.2,  FSharp.Core ≥ 6.0.0
 * new for `Expecto.Expect`